+++ src/cli/enum/Listing.ts
@@ -2,4 +2,5 @@ export enum Listing {
     ALL,
     DEPLOYED,
     LOCAL,
+    ADDABLE, // Guild files not deployed in the guild yet
 }